% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{boto3}
\alias{boto3}
\title{boto}
\usage{
boto3()
}
\value{
A reticulate module
}
\description{
\code{boto3()} is equivalent to \verb{import boto3} in Python
}
\details{
Gain Access to the boto3 Module

This is the base function that the entire \code{biggr2} package
runs off of. This is a wrapper for \code{reticulate::import('boto3')}
}
\examples{
\dontrun{
boto <- boto3()
}
}
\concept{Python Module}
