% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boto3.R, R/client.R
\name{boto3}
\alias{boto3}
\title{boto}
\usage{
boto3()

boto3()
}
\value{
A reticulate module
}
\description{
\code{boto3()} is equivalent to \verb{import boto3} in Python

These are the latest documents:
\itemize{
\item \href{https://boto3.amazonaws.com/v1/documentation/api/latest/index.html}{boto3}
}
}
\details{
Gain Access to the boto3 Module

This is the base function that the entire \code{biggr2} package
runs off of. This is a wrapper for \code{reticulate::import('boto3')}
}
\examples{
\dontrun{
boto <- boto3()
}
}
\concept{python}
