% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{client}
\alias{client}
\title{client}
\usage{
client(
  service = NULL,
  key_access = Sys.getenv("AWS_ACCESS"),
  key_secret = Sys.getenv("AWS_SECRET"),
  region = Sys.getenv("AWS_REGION")
)
}
\arguments{
\item{service}{NULL An AWS Service like 'ec2' or 's3'}

\item{key_access}{NULL Your AWS ACCESS Key}

\item{key_secret}{NULL Your AWS Secret Key}

\item{region}{NULL Your preferred AWS Region, mine is us-east-2}
}
\description{
Client Level Access to AWS Services
}
\details{
Access an AWS Client Service
}
\concept{client}
