% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_list_buckets}
\alias{s3_list_buckets}
\title{Print S3 Bucket Information}
\usage{
s3_list_buckets(s3_client = NULL)
}
\arguments{
\item{s3_client}{A connection to a S3 Client}
}
\value{
A tibble
}
\description{
Print S3 Bucket Information
}
\examples{

\dontrun{
# Return Information About Existing S3 Buckets
s3_list_buckets(
  key_access = Sys.getenv("AWS_ACCESS"),
  key_secret = Sys.getenv("AWS_SECRET"),
  region = Sys.getenv("AWS_REGION")
)
}
}
\concept{describe}
\concept{s3}
