% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_list_buckets.R
\name{s3_list_buckets}
\alias{s3_list_buckets}
\title{s3_list_buckets}
\usage{
s3_list_buckets(client = NULL)
}
\arguments{
\item{client}{A connection to a boto3, S3 Client}
}
\value{
A tibble
}
\description{
\code{s3_list_buckets} returns a data frame of bucket names and their creation date

\href{https://boto3.amazonaws.com/v1/documentation/api/latest/guide/s3-example-creating-buckets.html}{s3-example-creating-buckets}
}
\examples{
\dontrun{
library(biggr2)

s3_client <-
  client(
    service = "s3",
    key_access = Sys.getenv("AWS_ACCESS"),
    key_secret = Sys.getenv("AWS_SECRET"),
    region = "us-east-2"
  )

s3_list_buckets(client = s3_client)
}
}
\seealso{
Other s3: 
\code{\link{s3_create_bucket}()},
\code{\link{s3_delete_bucket}()}
}
\concept{s3}
